Action Frame.
Added the action frame which is the main window of the app. When you login as an it, the login window disappears and then the action frame appears. Added a menu bar with all the choices that the exercise asks for. The buttons of the menu bar are yet to be programmed.