GUI changes.
Added two new classes: ConnectToMySQL that uses the method getConnection to connect to the database and BranchInfo that it was mostly for testing and to understand some things. I'll go by that approach since it was successful and for each choice that the user makes in the LoginFrame a new frame with a table based on the choice will appear.