Gui and stored procedure updates.
Added the customer info frame that searches for customers that have the same lastname and then it prints their names and the offer id of the trip on the first table. On the second table it prints how many customers with the same lastname are there in each offer. In the stored_procedures.sql i added one extra stored procedure that's called find_lastname_count that basically splits the data from the 2 different tables into two different procedures.